Remove unnecessary Canvas2D story, refactor DatasetItemWorkspace story
